File reorg and dynamic configuration
Move files around to clearly set the configuration(s) apart from the EP.
As a consequence one must force rebuilt the configuration whenever the
EP is changed (fixable).

Also, made configuration a function that can thus be called repeatedly
when the underlayig DB changes. The live update of the stream processor is
yet to be implemented, though.