Allow select clauses to use aliases declared earlier.
This used not to compile:

SELECT max thing, max_thing+1

This should have raised an error at submission but didn't because we
automatically look for fields with no explicit tuple prefix in out when
they cannot be found in in.

Then it raised an error at compilation time, since the expression that
build the out tuple cannot refer to out tuple fields.

Instead of disallowing this convenient pattern, this patch changes the
generated code for the emission of the out tuple so that it can refers to
previously computed fields under their normal name (ie as fields of the
out tuple).
Also, allow the use of the out tuple in all select clauses, despite the
compilation would still fail if we refer to a field that is only defined
later in the select clause. Will have to add a check for this later. This
whole tuple usage check is not correct anyway and have to be redesigned
entirely.