Implement the EXPORT keyword
SELECT AND EXPORT will add ramen itself as a "children" of that node, and
for now ramen will merely read all those additional ringbufs without
paying attention to what's in there.

Later: we want to save the last N tuples of each exporting nodes, and make
them available in json via a websocket.

Ref #15.