Use signature to name generated code source files and binary
Which save on building the same binaries, and also allows the OS to share
the text segment of those binaries.

In cases where we have plenty of identiqual operaitons, this has almost the
same effect as having a single binary (#22). Implementation of parameters would
help even more (as it would remove unsubstancial variations of in litteral
values from the operation).