Demo of persistant state for percentiles
Here the 'Persistent' module is used for the percentile operation with
some hardcoded instructions about when to save on disc.

This is not satisfactory yet because we need the operation to specify
which kind of persistance is needed based on the expected number of
states and volatility of keys. Each node should come with a prefered
persistant strategy and we should shoose the module according to it.

Ref. #35