Get rid of Select expression altogether.
Now all select operations are actually aggregates.
Runtime cost is low (a bit of extra book-keeping that we could prune
if needed) but it's simpler as we now save all state in one go.
Main annoyance is we always have this large hash table that would be empty
for pure selects (but still large due to the large number of buckets),
and we are going to save/restore it regularly for nothing. Very easy to
optimize away in the future shall we need it.

Closes #38