Transition toward generalized stateful functions, step 1
Allow aggregate functions within aggregate functions (as in fun_with_funcs
test).

Mostly, inverted the depth order of Expr.iter_fun and turn the group state
initialization function from:

  let init tuple = { field1 = value1 ; field2 = value2 ; ... }

into:

  let init tuple =
    let field1 = value1 in
    let field2 = value2 in
    { field1 ; field2 ; ... }

but ensuring that the fields are evaluated in depth first order, and that
the code emission for expressions of value1/value2 would not get the
states from the aggr_ record (still to be created) but from the current
environment (using a inelegant ~state partly hidden parameter).

Also started to rename some of the aggr_ into state or group_state, and
added some comments.