Function `hash` to turn anything into an i64.
In conjunction with `remember` allows to remember combinations of several
fields without any need to make remember variadic.

Note: a variadic remember would be complex since arguments would have any
type ; but we currently transmit variadic arguments in an array. We would
have to build a custom tuple for that, requiring mode code generation. A
hash function is good enough.

Closes #56