Fix CASE typing.
Especially: we used to copy the type of a selected field expression when
adding it to the output type vector. This was necessary because we could
not afford to alias them with input types (`SELECT foo` would have aliased
the input type of foo and then updated it when modifying the expected
output type). I think that was the only reason behind that. The problem is
that we do not have in the expression the result of the typing for the
first level of expressions (which is typed in the out vector).
To fix that, we now copy the type only when the expression is a field;
otherwise we happily alias the type of the output vector with the type of
the expression. It seems to work.

Also, we still cannot find out the nullability of CASE expressions.

Ref #58