Add IF and COALESCE
IF is just syntactic sugar for CASE, but COALESCE is actually an
expression of its own. Since CASE does not allow to get rid of some
NULLability. consider:

  CASE WHEN foo IS NOT NULL THEN foo
       ELSE 42
  END

This expression would return a nullable integer.
It require a good deal of analysis to find out if it's Ok to set a CASE
not nullable (on more complex cases than the one above) and we can't be
certain that's actually what the user wants.

Therefore, COALESCE will be used for this: to make a value NOT NULL by
providing it a default value (or several). To achieve this, we have to
restrict its use a bit compared to SQL: The last alternative of a COALESCE
must not be nullable (like a constant) and all other alternatives MUST be
nullable. Then, the result is not nullable.