GUI: simpler and more robust way to find out a Fun node must be redrawn
Bug was: if a change was made to a node that was not visible (like a
change in the layer editor that we save at the same time we cancel the
form) then the info that this param changed will be lost by the time it
is made visible again, and therefore the old version of the DOM will be
displayed.
With this change it is much simpler: we merely save the "time" at which
each param changed for the last time, and the time each Fun DOM was
refreshed for the last time, and compare the two.