Fix: do not keep history when a node output type is changed
Bug was: crash when reading old history files after a node has been edited.

In general we probably want to keep the history as much as possible.
So by default, we keep it until the output type changed (aka the directory
it's stored into depends on the hash of the output type).
Also, this dir uses the node name so renaming a node effectively prune the
history.
Beside that, history is kept.