Fix: do not let a where clause uses any out tuple.
We do not pass any out to where_slow for a good reason: we wouldn't be able
to call it at the beginning or when creating a new group.

Even if we would compute the out tuple before calling where_slow, which would
at least provide a tuple of the proper type to where, if not the expected one,
at the beginning, we would now have the problem that the tuple generation
function also updates the group state, that we don't want to update when the
where filter fails.