Implement `COMMIT BEFORE`
`COMMIT AFTER` is now synonymous for `COMMIT WHEN`.

`COMMIT BEFORE` will output the tuple *before* the new incoming tuple is
merged in. It is useful when it's easier to detect the first tuple of the
next window than the last tuple of the current window.

Still to do: make it work when the commit condition applies to all groups.

Closes #141