Fix ramen_configurator for BCNs
prefix_smart think that zone_src and zone_dst, which fields are not
mentionned but as out fields, belongs to out even in a where clause
where the out tuple is not allowed. Therefore, as a quick fix, use
an explicit "in" prefix.

Alternative: in where clause do never assume a field comes from out?