Add time information in instrumentation and use it to display rates in GUI
The GUI part is merely a hack just OK to display rates of interesting
values. Ideally we'd like to get rid of the whole top-nodes panel and make
an internal data source from the instrumentation instead, so we could
compute rates using the DSL and display it as any other data stream.

Closes #93