Fix deadloop while outputing tuples
Bug was that when a child is removed when its in ringbuf is mostly full
then its parent may end up trying to overfill it before checking its
out_ref again.
Fix is: retry to output only if the file i still in the out_ref list.

This should probably be done only once every N failures.

Closes #165