Make it possible to edit running layer
Using a special flag `ok_if_running` in the put_layer RPC, set to on for
ramen_configurator.

Still TODO: make the start/stop of process atomic; for now if the
compilation fails the old configuration will be preserved but formerly
running processes will still be stopped. See FIXME added in HttpSrv.ml.

Closes #31