hysteresis helper functions
One can now write:

  hysteresis(previous_value, new_value, min, max)

or

  hysteresis_min(previous_value, new_value, threshold)
  hysteresis_max(previous_value, new_value, threshold)

and the result will be true as soon as the value venture outside the
allowed range, and will return to false only after the value is back into
the range, at least 15% within the boundary (an optional last parameter
allows to change this 0.15 into some other ratio).

Notes:
- previous_value must not be NULL (or we wouldn't know in which
  state to start);
- This is just syntactic sugar.

Interrestingly, this required to add a small improvement to typing.
Consider:

  SELECT group.previous.x > 0 as x

From this we can infer that x scalar type must be boolean. But can it also
be null? Nothing tells us, and in these cases we can have it both ways.
Of course the typer prefer to have it non-null.

Closes #75