Fix crash when reading instrumentation from a worker
The reason is that we read the instrumentation ringbuf using
RingBuf.read_next, which does not need the rb but only the tx (which
mvalue has a copy of the rb pointer). Therefore the rb mvalue is not
needed any longer and the GC reclaim it. Problem is: its finalizer unmap
the ringbuf.

Automatic memory management can be tricky.
I don't remember relying on the finalizer to unload the rb so let's get
rid of it entirely.

Closes #245