`ramen graphite`: fix expansion in case a factor has no value
It used to be that expanding a query stopped at the factor level when a
factor field had no value yet (cartesian product of possible values with
field names), which was especially confusing from Grafana. Now we add a
fake empty string for unmanned factors, so that the iteration proceed
towards the tree leaves. Grafana conveniently ommit those empty strings,
only offering to autocomplete with an "*", which is nive.

Closes #253