Fix: prevent TOP/global state update mix known not to work
With TOPs it is not that simple that global updates will be performed only
once per input tuple (which is the assumption I guess). In particular,
in `TOP by X WHEN Y` Y is not tested when we end up without a group (when
the input tuple end up in the "other" category).