`ramen supervisor`: dummy mode
Implemented as a general "experiment" feature sllowing to branch the
code according to some setting that depends on the site.

The first experiment is the "dummy mode", which is the default (since
it's supposed to be the null hypothesis).

The statistics and metric gathering are still missing but at least the
mechanism to enable/disable the dummy mode are done properly (famous
last words).

So from now on: `ramen supervisor --variant=TheBigOne=on` or

export RAMEN_VARIANTS=TheBigOne=on

Closes #167