Simplify (and fix) typing of NULLs
Get rid of the NULL type (but not of the NULL value).
The "NULL" literal is now of _no_ type (to be determined by the typer).

While at it, also fixed the error-prone design choice that the COALESCE
operator should turn into the identity function when it has only one
argument.