WIP: do not use Arrays for tuples
As actual solvers do not support mixing arrays and datatypes in
practice.

We introduce additional variables (such as "e12_2" for the 3rd element
of tuple "e12") but only some of them will be featured in the output.
The problem is that we may assert that e12 is the same as some tuple
(from the parent for instance), and that e12_2 is a string, with no
information on e12_0 and e12_1 because it's not needed and we don't even
know the number of elements in the tuple. Of course, we must report
missing types from parent tuples and also choose every type that is not
set by the solver to whatever type.