SMT: Fix tuple typing
Using sub identifiers (eXX_YY) for tuple items do not work because of some
functions returning expressions that turn out to also be tuple, for
instance:

  1st (2nd (1, (42,2)))

2nd type would properly be identified as tuple "(1, (42,2))" second item
type (eXX_2) but 1st type would not be related to 2nd type.

So, instead, use a specific type for each tuple size, which seems to work
well.
TBD: gather the actual tuple sizes that are needed instead of relying on
hardcoded max size.