Finish implementation of lists
Lists are like vectors of variable length.
In the ocaml generated code, both are represented as Array.
Serialization of list is a bit harder due to the fact that size is only
known at run time.
Changed the type declarations so that "I32[]" is now a list, not a vector
of yet-unknown dimension.