Propagate nullability in casts (as before)
Since recent changes one has to write "string?(x)" is x is nullable
and "string(x)" is it is not. This is inconvenient for no good reason:
better treat casts as ordinary functions that propagates nullability.

Now, "string?(x)" would return a nullable x even when x was not
nullable, which might be useful.