Rework notification part1: make all parameters (and name) expressions
Still replace "${thing}" with the value of field thing, so that the ramen
source files are compatible.

Simplify notification and team configuration by getting rid of severity
(want deferrable notifications? Assign them to another team!).

Now have event time, sent time and receive time with every notification.
No need to pass a specific parameter for that.
Do not use PPP to pass the notification message but a dedicated tuple type.

Ref #273