Change NOTIFY parameters syntax to avoid ambiguous syntax with AND
Indeed, AND is now a delimiter for the list of parameters, for the list of
commit clauses, and also a valid operator within an expression. So that:

  WITH PARAMETERS "name" = value AND KEEP ALL

for instance would fail while parsing "value AND KEEP ALL" as the value of
the parameter, since the binary_op reducer of parsercombinators evaluates
the operator eagerly (not doing so would slow down the parser
significantly).

So instead, the syntax becomes:

  WITH value AS name AND KEEP ALL