Fix use of outdated caches after autoreload
The graphite API used to cache name expansion based on the rc mtime, and
probably other places to similar things. And to be honest, there is no
other ways to detect when the configuration changed.
So this patch make sure the process supervisor touches the rc file
whenever the running configuration is changed.

Closes #331