Limited form of relative program names in FROM clauses
At compile time, parents will be looked for in the root_path according
to the actual (or specified with -o option) path of the program being
compiled. At link time, parents will be looked for in the running config,
relative to the program name being linked (ie for now the onw set during
compilation with -o, but soon the mount point given to `ramen run`).

This allows for out-of-tree compilation of ramen programs, such as the
Zones.ramen example.

Ref #163