Limit the calls to function finalizers
Finalizers are now only called for fields accessed via group.previous,
via the commit condition or when their value is required to upgrade
the state of another field, such as in:

  SELECT 95th percentile(x) as FOO, AVG(FOO) AS BAR"

This makes functions using slow finalizers over long aggregates much faster.

The down side is that "COMMIT BEFORE" is no longer available. Indeed,
to emit the previous tuple is no longer possible as we don't compute it.

Possible solution:
- finalize all fields in case of COMMIT BEFORE, so that
  minimal_typ = out_typ;
- keep previous input tuple and all previous context required to compute
  the previous out_typ.

For now it's just disabled. Before doing anything the context has to be
cleaned, and all useless fields and counters removed.

Ref #356