Assert that no ringbuf record has length zero
The only legit(?) case of empty record were when a select operation
selected no public fields and merely notified. We now skip outputing a
tuple altogether when it's empty (we had no way to "select" nothing from
a hypothetical child anyway).
Since there is a deserialization that sometimes fails with record
size = 0 let's abort as soon as such a record is written.