Stop hotfixing multi-reader/writer ringbufs
We used to detect when a concurrent writer died and delete its uncommitted
record instead of freezing forever.
Unfortunately, this worked only when _all_ other concurrent writers died.
In case of several still alive writers attempted to hotfix the ringbuffer,
corruption would ensue.

The fix is to not do anything and accept to stuck frozen and be killed by
a watchdog, then restarted by the supervisor, which also might clean the
ringbuffer.

Depends on #401 to work properly though.

Closes #398
Closes #378