`DEFINE IF _some_condition_here_ ...`
With this syntax it is possible to control at runtime whether a worker
should run or not. If not, it will be featured in `ramen ps` list but
will perform no work at all (actually, no process will be running).
Regarding linking, it's as this function were not running (which it isn't,
really).

Closes #492