Have a single run condition per program and test it at `ramen run` time.
The problem with running and then allowing to stop a function that was
disabled by experiment, is that it was still present in its parent out_ref
and therefore blocking them.
Now we interpret the running condition more literaly: when a program must
not run, it just isn't added to the RC config. It's ok because `ramen run`
does not then return an error.
We can no longer have a different running condition per function though,
as we run programs as a single unit.

Closes #495