Fix aggregates on lists
It is mandatory to reset the state of the aggr function for when we KEEP
ALL.

This fix makes the whole code pretty awful though.

One option would be to turn those stateful functions into stateless ones
right after typing (and warn if the state has been changed from default).
But that would not be great either.
We should maybe have an helper function `aggregate_on_list` that would
prevent the generation of a global/group state for the function state
and instead we would use a local state at the point where the aggregation
is compiled ; meaning we have to pass the state as a special local variable
instead of a state variable for the compilation of the actual aggregation
to use that local variable instead of whatever state is specified in the
[g] parameter. Or, wy not have directly a third possible state alongside
group and global: local, for that case, and we would put it automatically
after typing (also checking it's still None).