Do not archive new workers by default
Archiving at start was a way to give time for an archival configuration
to arrive (from a first tail or timeseries via graphite for instance).
This was also convenient for debugging. But the cost in IO for high
throughput nodes is just too high (we should be able to set the expiry
via timeout or number of written tuples). Typical scenario: ramen is
started with a load of CSV already waiting, and because of the archival
it takes 10 minutes longer than required to catch up.

It seams thus safer to not archive by default.