Default retention should be to store only the roots
Only for 10 mins so that it's very likely to fit within 500Mb (default
allocated storage size)

FIXME: actually it's rather stupid to specify absolute disk spaces.
Rather specify the cost of a saved byte and the cost of saving a byte and
rereading it. And then we could _warn_ about the total allocated space
being too small.

Closes #617