Enough with private fields!
Given we go at length to make sure only used fields are transmitted from
producer to cosumer worker, the only benefit of private fields is to save
HD space when archiving output. Unfortunately, it also makes typing of
the output tuple more complicated, and require a lot of special code
already.

A better solution to save storage would be to optimize away fields that
are computed solely with stateless functions from other output fields
(and recompute them automatically when asked for).

That does not eliminate many cases of private fields though...