ORC: An (untested) experiment to archive in ORC rather than RingBuf
As in the long term archival format might change, care must be taken to
share the same archival directory between the format, and that the tuple
iterator reading the archive can read both formats.

Even in ORC mode, non-wrapping ringbuf can also appear in many
circumstance, and that's a good thing as they can be read in flight while
ORC files cannot. So in practice archiving workers should still write in
a non-wrapping RB the "current" tupples before the ORC file appear in
the archive directory (beware of the interstices though; TBD).
So to avoid that these ringbufs are copied in the archive directory a new
flag is added to the ringbuf headers (which does not change the format
though).

Ref #667