Optimise removal of entries from RamenHeaps.
Also measure separately the time required to select the list of tuples
to commit from the commit itself.

Turns out, most of the time in tests/storage/out is spent outputing the
selected tuples, not selecting them and not removing them from the hash
and heap (as evidenced by separate time measurements).
Not serializing the output (these nodes have no children, they merely
archive) but computing the out_tuple from the min_out_tuple (which
performs a substring and 6 sampling finalization + percentile
computations).

So here, higher cardinality implies more output, which is slower roughly
in proportion.