Fix TX leakage
RingBufLib.read_ringbuf is a dangerous API which expect its callback
to free the TX (for perf reason) and does not check at the end that it
has been freed.

TODO: a flag in the TX cleared when the TX is dequeued and that
read_ringbuf could check after the callback had returned? We might not
even need a specific flag.