SELECT FROM site:function + major rewrite of supervisor for multi sites
Introduced `sites` and `services` instead of abusing `hosts`.
Command line options regarding `hostname` renamed to `site`.
Accordingly, changed service directory into a double hash (sites then
services).

Supervisor now build a whole representation of the multi-site layout to
decide what to run. This makes it easier to reason about, but is also
much slower. To be fixed later.

Also TBD: how should `ramen kill` and `ramen ps` deal with multi-site.

Note: it is no longer true that a FQ function name represents a unique
worker, as several workers implementing this function can now run on
various sites (and maybe one day on the same site) but it is still the
case that the FQ program name is a key in the RC file.

Closes #699