Make persistent keyword useful
It was initialy planed to set retention in the initial version of the
archivist, then removed from the archivist but not the syntax. Now it's
doing what I think it was doing, making the output to be archived
regardless of the retention config (but only up to when the GC decides
to do away with those files of course).

In the short term we want the whole retention config in there anyway, for
simplicity.