Avoid logging in signal handler to prevent deadlock
BatPrintf evenrtually uses BatInnerIO lock (a real lock since we compile
with threads). So that if the signal handler is executed while that lock
is already taken, then a deadlock would occur.

This has actually been frequently observed with the forked server killer
thread and the signal handler both blocked in printf.

For SIGUSR1 though, there is no way around it for now.