Fix bad use of advisory file locks
OCaml stdlib lockf is implemented with fnctl, which locks are released
as soon as any thread closes _any_file_descriptor_ to the file.

The code removing a replay from the replays file correctly takes an
exclusive lock, then read the replays, modify it, then write them back.
But to read them it used to reopen the file (as ppp_of_file took a file
name).

This adds ppp_of_fd and avoid reopening that file, thus not loosing the
lock.

Closes #748