Compilation via the confserver
First PoC that works:

While the confserver and the _compserver_ are running,
`ramen compile --confserver` will just get the compserver to *precompile*
the source and store the result in the conf tree.
Later `ramen supervisor` could get the precompiled program from there
and generate actual code, thus solving the sharing of binaries in
distributed settings.

Will now need to demo that in the GUI.