Fix receiving several time KValue->valueCreated signal
Problem was: when autoconnecting, some object might attach a listener
to valueCreated. As a convenience, autoconnect then emits the
valueCreated signal for every preexisting KValues matching the regexp.
It has to use `emit` so that the receiver will be called in the proper
thread. So, other receivers, which might have already been notified
of the new value, will also receive the valueChanged signal, unless
they disconnect right after that.

autoconnect cannot disconnect automatically the valueCreated due to what
look like an oversight from Qt (?) so each receiver has to do it.

Hopefully there is the Once library that makes this very simple and
saves the day.