Tail from a specific instance rather than just a specific site+fq
Still not perfect though, as workers do not remember previous run sequence
numbers. Those should be remembered (saved in their state? But then we
would have to force state save after every publication of a new tail,
while if it had its own dedicated file it would be lighter, and also
would survive state version change).

Large refactor of tailling from rmadmin which introduces a few new bugs
but the code is nonetheless clearer, so that count as a win.

Closes #893