rmadmin: attempt at expanding new things
Also make the proxy a Q_OBJECT (requiring to move it into its own .h so
that the preprocessor can find it), which does not improve the situation
but looks better.

Not good enough to expand unfiltered rows though.
But as I cannot see anything to achieve this short of reimplementing
a whole proxy filter, let's call it a day and...

Closes #896