rmadmin: coarser grained signalling for simpler code
and probably faster too.

Instead of connecting per values, connect once with the whole kvs (which
is now a mere std::map) and let the signaled filter themselves what they
are interrested into.
As a consequence, autoconnect is gone, the intrusive list is gone,
the once module is gone.
Almost none of the known bugs seems to be gone, though; but at least
tracking them will be easier.

Closes #911