compserver: save all intermediary targets in the compserver
Major overhaul of the internal builder.
Now info file preserve not only the md5 but also the extension of the
modified source that started the compilation chain.

Note: alerts with non existing parents still have no ramen file because
parents info are needed to get the field types and default aggregation
function for reaggregation.

Closes #919
Closes #920