Reworked `LAST` function
Now called `LARGEST` to account for the by-clause, or `LATEST` when
that clause is ommitted.
The main difference is the addition of the `UP TO` modifier that
determines what to do when fewer entries than specified are available.
Also, `SMALLEST` or `OLDEST` will reverse the selection order.

rmadmin: rename last function according to recent change

Fix new last names