rmadmin: simplify charts by getting rid of ChartDataSet
So that points belong to the same dataframe are handed together to the
Graphic instead of dealing with non-trivial race conditions. As a downside
it is no longer possible to have a chart with datasets from different
functions, but one could (and should) add a ramen function to (properly)
join the various tables into one so it's not a limitation in practice.

Also will make it easier to cache past data by frame rather that by fields.