RowBinary: use the new Dessser lib to generate the unserializer
The idea is that the Dessser lib is done in such a way as to permit the
progressive transition from OCaml to C code generation.

Next step is to fusion RamenTypes.t and RamenNullable with Dessser user
types.
Then to change CodeGen_OCaml internal representation of tuples to use
records instead, so that types become really interchangeable.
Then we would be able to use Dessser casually to generate any of the
functions needed by a worker. And when all those functions are doable with
the C backend, switch that worker to use C instead.