Fix compserver crash when asked to compile an orphan
When it receives a notification that a compilation have failed because
of a missing parent, the compserver performs a lookup of that parent to
attempt another compilation if the parent has shown up in the meantime.
While performing this second-chance look up it failed to handle the
Not_found exception result from the config tree and crashed miserably
each and every time the parent was still missing (which is almost
always).

Closes #1005