Fix rejecting z3 fantasies
In some cases the type-checker is allowed to make up types for unused
values, up to records with invalid field names.

Also simplify the test that was actually not legit and would force the
backend to have a converter from any value to any value, accepting only
to convert NULL values into NULL values and crashing at runtime for
anything else. It's much more useful to keep the compile time error
instead.

Closes #1030