Improve behavior when a RB contributor fails to update the tail pointer
Before:
If several readers (resp. writers) are contributing to the same RB, and
one of them fails while holding a TX either in the cons or prod area,
then all other contributors wait forever that the area tail advances.

After:
After some long time spinning, blocked readers/writers give up and proceed.
If there are several blocked processes that give up at essentially the same
time, there are chances that this creates an awful mess though
(essentially, the writer which tried to write last may be the first to
update the tail, jumping over the other writers, which will then also
overrwrite the tail back to where they think it should be, etc.
To avoid this, in that case instead of giving up and setting the tail where
the had was seen, each blocked processes slowly _increases_ the tail
pointer, so that as soon as it reaches the position seen by the first blocked
process everything goes back to normal.

Of course, the reader (in the case of a dead writer) will still have to
confront the likely erroneous record in the RB, which can also create quite
a lot of damage.

Ref #1031