Fix deserialization of vectors
Once a vector had been unserialized, following fields are likely
to cause an unserialization error.

Observed on ipv6-as-array-of-chars datatypes, but reproduced easily
in tests/func/constructed_types.ramen.

Actually the bug was a good old evaluation-order bug in the generated
code causing TX offset to be wrong after any vector was unserialized.

Closes #1043