Fix race condition when restarting a worker leading to unlinked children
When supervisor restart a function (either using the same worker, or
another instance with the same input type), then the former and the new
worker instances will compete to access the same input ringbuf (which
name depends only on the function input type by design). Since supervisor
now is asynchronous for handling waitpid, it frequently start the new
before learning about the killed former. Thus, when it unlinks the old
worker from its parents it will also unlink the new one (since those links
are not refcounted).

The fix adds the pid of the current owner of a link into the out_ref file,
and removes does nothing if the current owner is not the removed pid.

Closes #1053
Closes #251