Fix httpd resetting target_config for nothing
RamenRun.do_run do not reset the same RC entry ever again, regardless
of the `replace` option, which is now gone.
This was useful when program names where arbitrary but now a program
is automatically restarted whenever the source is changed, and the replace
flag is therefore useless.
do_run is now a nop when the same entry is already in the RC.

Closes #982