Fix deadloop of internal services when confserver is restarted
Scenario: confserver is restarted but other services keep going.
confserver forget all identities, zero-mq reconnects in the background all
clients but those are unaware that the connection have been interrupted
and keep expecting answers as if they were authenticated.
But the confserver cannot send any error messages back to them as they
are now anonymous (therefore without an error file).

The fix is simple: the only command that an anonymous client is allowed
to perform is Auth. Everything else is an AuthErr (which is delivered to
anonymous clients). At reception, clients fail and restart.

Closes #1079