Fix supervisor whining about missing pids
The error case was not supposed to ever happen, yet it did.
The reason for that weird behavior seems to be that some unidentified
library explicitly ignores sigchild, which result in waitpid to err out
when a child exit instead of returning it's exit status.

This patch not only try to improve the behavior when this happen, but
also reset the sigchld handler to the default one at startup, which seems
to fix the problem.

Could find the guilty lib (looked at zmq and librdkafka source code but
could find anything there, maybe in one of their lib.