Store all source md5 sums in the precompiled info
Required because the builder might skip input files if they do not need
to be rebuild without the client being aware of it. Also the current logic
was not implemented properly anyway due to the fact that precompiler would
only ever perform the first step of every build chains (relying on further
key update from the confserver to trigger the next steps). That's nice and
robust in some sense, but it was then quite hard to track what the original
source change triggered the build.

This patch revert this behavior to a simpler one where the whole chain is
performed in a single call (and ignoring all received source changes when
authored by the compserver itself), thus keeping track of all involved
md5s properly.

This is unfortunately less robust in case the compserver dies in the
middle of the chain.