Test alerts
Make every workers able to optionally send test alerts regularly.
Test alerts are made of real notification settings, but triggers
based on time independently of actual condition.

A test alert therefore has real values and is routed and formatted like
the real alerts that the worker would fire, but has an additional "test"
label that can be used to ignore them.

This is useful both for debugging/testing alert delivery but also to
make certain once in a while that the expected alerts are actually
running.

Closes #1125