Make alerts-v1 able to track several groups in a single worker
Case in point: define an alert with same thresholds etc over field X of
some function grouped-by Y and Z, with filter Y in [1,2,3] and Z=42,
meaning: alert _individualy_ for each of those Ys.
The equal condition on Z cancels the group-by Z, and the tuples must not
be reagregated. Rather, we want to GROUP BY Y, and alert independently
for each group.

Previously, we would have re-aggregated all values for all selected Y, and
have a single alert on the aggregate.