An attempt to go after #1206 but not satisfying
The problem is that breaks the assumption that the only writer to info is
precompiler. To make this patch we would have to make execompiler write
into info "silently", without changing the mtime. Also, the execompil
errors should be kept by site.
This just shows that this information does not belong here.
And indeed, execompil errors can only arise in case of codegen bug, and
does not indicate a problem with the precompiled info per se.

So, parked waiting for better ideas.