`ramen run --replace`
An option to force replacement of programs with same name.

Note that this also fix a bug: if a program with same name but different
params (or options) was already running, the new RC entry would be added
to the target_config that's an assoc-list with the former one left on that
list. Now either the new one replace the old one or the old one is kept
and the new one discarded with an error message.

Closes #1215