alerter: do not contact oncall when timeouting an alert
Make timing out a notification an internal process to clean the task
stack. Do not send a bogus message to oncall that the incident
recovered, since it has not (and alerter lacks informations to build a
proper recovery message anyway, leading to a misleading _firing_ message
instead in some cases).