Fix type-checking for NULLs and others
Using an int sort for integer widths meant that the solver had no
boundary for the width when no constraint was set on the type (such as
for plain NULL literal), leading to all kind of weird resul when trying
to minimize the total width.
Fixed by using a bitvector for the integer width. Should be faster as
well.