fix: ramen test should wait worker stats
This patch adds a way to wait worker stats in ramen test command.
It does this by sending a sigterm instead of sigkill to workers to
let them term properly.

Moreover, it wait the worker zmq thread to send all commands to
the confserver.

Finally, it wait the stats in the ramen test command by comparing
the time of the received RuntimmeStats and the time where
workers was killed.

fix #1254