Convert CSV used in tests to much simpler formatting for Dessser user types
Although it is possible to pass custom parser/pretty-printer for Dessser's
user types, it is not possible to give custom converter from/to strings
that could be used by the CSV (des)serializer to allow friendlier
representation of values in CSV files. So for now let's just serialize
them as integers.

Notice that dessser CSV (des)serializer does not even allow for hexadecimal
representation of integers. Lots of room for improvement there.