RaQL: change records syntax for the familiar '{ name: value; etc }'
The previous syntax was that of the SELECT clause, ie 'value AS name, etc'
but it was needed to disambiguate parenthesis because parsing
'(some complex parenthesized expression)' required to disprove tuple and
record in addition to trying the mere parenthesized expression.

This simple change make some patological cases more than 50 times faster.

Soon: rework tuple parsing to gain even more.