Do away with old ringbuf based stats/notif listeners
We want to refactor CodeGen_Ocaml to extract what's not specific to
OCaml in there (starting with the code to determine the minimual tuple).
Eventually, we are going to have a single operation: Aggregate ; so
that:
1. CodeGen_OCaml is reduced to generating some functions;
2. We benefit immediately from merging the first 2 layers of workers.

Selecting from stats or notifs will do a come back at some point in the
future, when the confserver values are encoded in a giant sum type by
libdessser and RaQL can deconstruct sum types.  Then, it will become
easy to just select those config values from the confserver that match
some key, and then pick whatever stats we want from whatever conf value.